+++ src/Components/Analyzers/src/Microsoft.AspNetCore.Components.Analyzers.csproj
@@ -20,7 +20,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="$(ComponentsSharedSourceRoot)src\ComponentsApi.cs" LinkBase="Shared" />
     <Compile Update="Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
