+++ src/Components/Blazor/Blazor/src/Hosting/ConventionBasedStartup.cs
@@ -9,7 +9,7 @@ using System.Runtime.ExceptionServices;
 using Microsoft.AspNetCore.Components.Builder;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.AspNetCore.Components.Hosting
+namespace Microsoft.AspNetCore.Blazor.Hosting
 {
     // Keeping this simple for now to focus on predictable and reasonable behaviors.
     // Startup in WebHost supports lots of things we don't yet support, and some we
