+++ src/Components/Blazor/Server/src/AutoRebuild/IRebuildService.cs
@@ -1,10 +1,10 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNetCore.Components.Server.AutoRebuild
+namespace Microsoft.AspNetCore.Blazor.Server.AutoRebuild
 {
     /// <summary>
     /// Represents a mechanism for rebuilding a .NET project. For example, it
