+++ src/Components/Blazor/Server/src/AutoRebuild/VSForWindowsRebuildService.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -7,7 +7,7 @@ using System.IO.Pipes;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNetCore.Components.Server.AutoRebuild
+namespace Microsoft.AspNetCore.Blazor.Server.AutoRebuild
 {
     /// <summary>
     /// Finds the VS process that launched this app process (if any), and uses
