+++ src/Components/Shared/src/CacheHeaderSettings.cs
@@ -1,0 +1,29 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.AspNetCore.Http;
+using Microsoft.Net.Http.Headers;
+
+namespace Microsoft.AspNetCore.StaticFiles
+{
+    internal static class CacheHeaderSettings
+    {
+        internal static void SetCacheHeaders(StaticFileResponseContext ctx)
+        {
+            // By setting "Cache-Control: no-cache", we're allowing the browser to store
+            // a cached copy of the response, but telling it that it must check with the
+            // server for modifications (based on Etag) before using that cached copy.
+            // Longer term, we should generate URLs based on content hashes (at least
+            // for published apps) so that the browser doesn't need to make any requests
+            // for unchanged files.
+            var headers = ctx.Context.Response.GetTypedHeaders();
+            if (headers.CacheControl == null)
+            {
+                headers.CacheControl = new CacheControlHeaderValue
+                {
+                    NoCache = true
+                };
+            }
+        }
+    }
+}
