+++ src/Components/test/testassets/BasicTestApp/Startup.cs
@@ -1,12 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Runtime.InteropServices;
 using Microsoft.AspNetCore.Blazor.Http;
-using Microsoft.AspNetCore.Blazor.Services;
 using Microsoft.AspNetCore.Components.Builder;
 using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Runtime.InteropServices;
 
 namespace BasicTestApp
 {
