+++ src/Components/test/testassets/ComponentsApp.Server/ComponentsApp.Server.csproj
@@ -11,6 +11,7 @@
   <ItemGroup>
     <Reference Include="Microsoft.AspNetCore.Components.Server" />
     <Reference Include="Microsoft.AspNetCore.Mvc" />
+    <Reference Include="Microsoft.AspNetCore.StaticFiles" />
     <ProjectReference Include="..\ComponentsApp.App\ComponentsApp.App.csproj" />
   </ItemGroup>
